+++ resources/views/tasks/show.blade.php
@@ -18,10 +18,14 @@
             <td>{{ $task->status }}</td>
         </tr>
     </table>
-    {!! link_to_route('tasks.edit', 'タスク編集', ['task' => $task->id], ['class' => 'btn btn-light']) !!}
     
-    {!! Form::model($task, ['route' => ['tasks.destroy', $task->id], 'method' => 'delete']) !!}
-        {!! Form::submit('削除', ['class' => 'btn btn-danger']) !!}
-    {!! Form::close() !!}
+        
+    @if (\Auth::id() === $task->user_id) 
+        {!! link_to_route('tasks.edit', 'タスク編集', ['task' => $task->id], ['class' => 'btn btn-light']) !!}
+        
+        {!! Form::model($task, ['route' => ['tasks.destroy', $task->id], 'method' => 'delete']) !!}
+            {!! Form::submit('削除', ['class' => 'btn btn-danger']) !!}
+        {!! Form::close() !!}
+    @endif
     
 @endsection
